sys/src/ape: use uintptr_t (thanks Geoff Collyer)

+++ sys/src/ape/lib/fmt/dofmt.c
@@ -13,6 +13,7 @@
  */
 #include <stdarg.h>
 #include <string.h>
+#include <inttypes.h>
 #include "utf.h"
 #include "fmt.h"
 #include "fmtdef.h"
@@ -348,7 +349,7 @@ __ifmt(Fmt *f)
 		break;
 	}
 	if(f->r == 'p'){
-		u = (ulong)va_arg(f->args, void*);
+		u = (uintptr_t)va_arg(f->args, void*);
 		f->r = 'x';
 		fl |= FmtUnsigned;
 	}else if(fl & FmtVLong){
